unloggedin user will be direct to signup page

+++ frontend/src/components/Tests/TestOne.js
@@ -1,16 +1,23 @@
 import React, { useState } from 'react';
 import test1Image from './TestOne.jpg';
 import { addResultThunk } from '../../store/result';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { useHistory } from 'react-router-dom';
+import LoginFormModal from '../LoginFormModal'
 
-const TestOne = () => {
+const TestOne = ({ isLoaded }) => {
     const dispatch = useDispatch()
     const history = useHistory()
     const [option, setOption] = useState(0)
     const options = Array.from({ length: 21 }, (_, index) => index + 1);
+    const sessionUser = useSelector(state => state.session.user);
+
     const handleSubmit = async (e) => {
         e.preventDefault();
+        if (isLoaded && !sessionUser) {
+            return history.push('/signup')
+        }
+
         let result;
         if ([1, 3, 6, 7].includes(parseInt(option))) {
             result = `Positive and enthusiastic are your strength! You will not be easily defeated by any obstacles. You firmly believed that "What does not kill you will make you stronger". As long as you keep your positivity, you and overcome anything. You are every likely to be successful`
